moved gui_components into src
