--- gui_components/__init__.py
@@ -1,9 +1,0 @@
-import gui_components.menu.main as menu
-import gui_components.view.main as view
-import gui_components.edit.main as edit
-import gui_components.settings.main as settings
-
-Menu = menu.Menu
-View = view.View
-Edit = edit.Edit
-Settings = settings.Settings
