+++ src/gui_components/__init__.py
@@ -1,0 +1,9 @@
+import src.gui_components.menu.main as menu
+import src.gui_components.view.main as view
+import src.gui_components.edit.main as edit
+import src.gui_components.settings.main as settings
+
+Menu = menu.Menu
+View = view.View
+Edit = edit.Edit
+Settings = settings.Settings
