+++ src/gui_components/edit/commands.py
@@ -2,7 +2,7 @@ import tkinter as tk
 
 from src.common import config
 from src.routine.components import Point
-from gui_components.interfaces import Frame
+from src.gui_components.interfaces import Frame
 
 
 class Commands(Frame):
